+++ checkov/terraform/graph_builder/foreach/abstract_handler.py
@@ -1,15 +1,19 @@
 from __future__ import annotations
 
 import abc
+import json
+import re
 import typing
 from copy import deepcopy
-from typing import Any
+from typing import Any, Optional
 
 from checkov.common.util.consts import RESOLVED_MODULE_ENTRY_NAME
 from checkov.terraform import TFModule
-from checkov.terraform.graph_builder.foreach.consts import COUNT_STRING, FOREACH_STRING, COUNT_KEY, EACH_VALUE, EACH_KEY
+from checkov.terraform.graph_builder.foreach.consts import COUNT_STRING, FOREACH_STRING, COUNT_KEY, EACH_VALUE, \
+    EACH_KEY, REFERENCES_VALUES
 from checkov.terraform.graph_builder.graph_components.block_types import BlockType
 from checkov.terraform.graph_builder.graph_components.blocks import TerraformBlock
+from checkov.terraform.graph_builder.variable_rendering.evaluate_terraform import evaluate_terraform
 from checkov.terraform.graph_builder.variable_rendering.renderer import TerraformVariableRenderer
 
 if typing.TYPE_CHECKING:
@@ -169,3 +173,74 @@ class ForeachAbstractHandler:
             tf_moudle: TFModule = config[RESOLVED_MODULE_ENTRY_NAME][0].tf_source_modules
             ForeachAbstractHandler._update_nested_tf_module_foreach_idx(original_foreach_or_count_key, original_module_key,
                                                                         tf_moudle)
+
+    def _handle_static_statement(self, block_index: int, sub_graph: Optional[TerraformLocalGraph] = None) -> \
+            Optional[list[str] | dict[str, Any] | int]:
+        attrs = self.local_graph.vertices[block_index].attributes if not sub_graph \
+            else sub_graph.vertices[block_index].attributes
+        foreach_statement = attrs.get(FOREACH_STRING)
+        count_statement = attrs.get(COUNT_STRING)
+        if foreach_statement:
+            return self._handle_static_foreach_statement(foreach_statement)
+        if count_statement:
+            return self._handle_static_count_statement(count_statement)
+        return
+
+
+    def _handle_static_foreach_statement(self, statement: list[str] | dict[str, Any]) -> Optional[list[str] | dict[str, Any]]:
+        if isinstance(statement, list):
+            statement = self.extract_from_list(statement)
+        evaluated_statement = evaluate_terraform(statement)
+        if isinstance(evaluated_statement, str):
+            try:
+                evaluated_statement = json.loads(evaluated_statement)
+            except ValueError:
+                pass
+        if isinstance(evaluated_statement, set):
+            evaluated_statement = list(evaluated_statement)
+        if isinstance(evaluated_statement, (dict, list)) and all(isinstance(val, str) for val in evaluated_statement):
+            return evaluated_statement
+        return
+
+    def _handle_static_count_statement(self, statement: list[str] | int) -> Optional[int]:
+        if isinstance(statement, list):
+            statement = self.extract_from_list(statement)
+        evaluated_statement = evaluate_terraform(statement)
+        if isinstance(evaluated_statement, int):
+            return evaluated_statement
+        return
+
+    def _is_static_foreach_statement(self, statement: list[str] | dict[str, Any]) -> bool:
+        if isinstance(statement, list):
+            statement = self.extract_from_list(statement)
+        if isinstance(statement, str) and re.search(REFERENCES_VALUES, statement):
+            return False
+        if isinstance(statement, (list, dict)) and any([re.search(REFERENCES_VALUES, s) for s in statement]):
+            return False
+        return True
+
+    def _is_static_count_statement(self, statement: list[str] | int) -> bool:
+        if isinstance(statement, list):
+            statement = self.extract_from_list(statement)
+        if isinstance(statement, int):
+            return True
+        if isinstance(statement, str) and not re.search(REFERENCES_VALUES, statement):
+            return True
+        return False
+
+    def _is_static_statement(self, block_index: int, sub_graph: Optional[TerraformLocalGraph] = None) -> bool:
+        """
+        foreach statement can be list/map of strings or map, if its string we need to render it for sure.
+        """
+        block = self.local_graph.vertices[block_index] if not sub_graph else sub_graph.vertices[block_index]
+        foreach_statement = evaluate_terraform(block.attributes.get(FOREACH_STRING))
+        count_statement = evaluate_terraform(block.attributes.get(COUNT_STRING))
+        if foreach_statement:
+            return self._is_static_foreach_statement(foreach_statement)
+        if count_statement:
+            return self._is_static_count_statement(count_statement)
+        return False
+
+    @staticmethod
+    def extract_from_list(val: list[str] | list[int] | list[list[str]]) -> list[str] | list[int] | int | str:
+        return val[0] if len(val) == 1 and isinstance(val[0], (str, int, list)) else val
